Extract App class from AddonTestApp
