+++ lib/index.js
@@ -1,7 +1,9 @@
 'use strict';
 
+var App = require('./models/app');
 var AddonTestApp = require('./models/addon-test-app');
 
 module.exports = {
+  App: App,
   AddonTestApp: AddonTestApp
 };
