+++ lib/models/addon-test-app.js
@@ -1,16 +1,18 @@
 var fs         = require('fs-extra');
 var path       = require('path');
 var findup     = require('findup-sync');
-var startServer = require('../commands/start-server');
-var chdir      = require('../utilities/chdir');
+var util       = require('util');
+var App        = require('./app');
 var debug      = require('../utilities/debug');
-var runCommand = require('../utilities/run-command');
 var pristine   = require('../utilities/pristine');
 var RSVP = require('rsvp');
 var copy = RSVP.denodeify(require('cpr'));
 
 function AddonTestApp() {
+  App.call(this);
 }
+
+util.inherits(AddonTestApp, App);
 
 // Public API for putting an app under test. If the app doesn't
 // exist already, it will create it and put it into the `pristine`
@@ -32,55 +34,11 @@ AddonTestApp.prototype.create = function(appName, options) {
     });
 };
 
-AddonTestApp.prototype.run = function() {
-  var previousCwd = process.cwd();
-  chdir(this.path);
-
-  return runCommand.apply(null, arguments)
-    .finally(function() {
-      chdir(previousCwd);
-    });
-};
-
-AddonTestApp.prototype.runEmberCommand = function() {
-  var cliPath = path.join(this.path, 'node_modules', 'ember-cli', 'bin', 'ember');
-  var args = Array.prototype.slice.apply(arguments);
-
-  return this.run.apply(this, [cliPath].concat(args));
-};
-
-AddonTestApp.prototype.filePath = function(filePath) {
-  return path.join(this.path, filePath);
-};
-
 AddonTestApp.prototype.editPackageJSON = function(cb) {
   var packageJSONPath = path.join(this.path, 'package.json');
   var pkg = fs.readJsonSync(packageJSONPath);
   cb(pkg);
   fs.writeJsonSync(packageJSONPath, pkg);
-};
-
-AddonTestApp.prototype.startServer = function(options) {
-  var app = this;
-  var previousCwd = process.cwd();
-  process.chdir(app.path);
-
-  return startServer(options)
-    .then(function(server) {
-      app.server = server;
-      process.chdir(previousCwd);
-    });
-};
-
-AddonTestApp.prototype.stopServer = function() {
-  if (!this.server) {
-    throw new Error("You must call `startServer()` before calling `stopServer()`.");
-  }
-
-  this.server.kill('SIGINT');
-  this.server = null;
-
-  return RSVP.resolve();
 };
 
 function copyFixtureFiles(appName, destDir, fixturesPath) {
